refactor(#741): refactor rssController
